backend/OpenAI/requirements.txt
# requirements.txt
requests
openai
python-dotenv
